Pass model ref to getVal, pass selector to observe fns
One relatively small change (which unfortunately has the potential to be breaking).  It's just passing a reference to the model to `getVal` methods in handlers.  Unfortunately it's been so long since I made that change in our local copy, I don't remember the exact use case we needed it for, but here's a potential use case (that is unfortunately not nearly as good as the actual use case we wanted this feature for in the first place.) I know there are better ways to accomplish this particular use case, but our original problem wasn't so easy to solve in a different way, so this is really more of an example than a pressing use case:

If this is the `$el` specified in the binding config:
```js
<select name="processAs" data-lookup="currentUserPerms">
    <option value="0">System Admin</option>
    <option value="1">Group Admin</option>
    <option value="2">Local Admin</option>
    <option value="3">This User Only</option>
</select>
```
And in our model attributes we had:
```js
currentUserPerms = [1, 2, 3]
```
Then in the binding config we could do something like this:
```js
getVal: function ($el, e, conf, model) {
    var enumAttr = $el.data('lookup');
    var perms = model.get(enumAttr);
    var potentialVal = $el.val();
    if (_(perms).contains(potentialVal)) {
        return potentialVal;
    } else {
        $el.trigger('permissions:invalid');
        return false;
    }
},
updateModel: function (val) {
    return (val !== false);
}
```